requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.3.3
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langsmith==0.1.142
numpy==1.26.4
openai==1.54.3
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.40.1
tabulate==0.9.0
Unidecode==1.3.8

